+++ packages/common/log/src/processors/index.ts
@@ -5,4 +5,4 @@
 export * from './console-processor';
 export * from './debug-processor';
 export * from './browser-processor';
-export * from './json-processor';
+export * from './file-processor';
